removed redundant cast

+++ src/number-of-1-bits/ans.go
@@ -4,7 +4,7 @@ func hammingWeight(num uint32) int {
 	var count uint32 = 0
 	var i uint32
 	for i = 0; i < 32; i++ {
-		count += uint32(num>>i) & uint32(1)
+		count += num >> i & uint32(1)
 	}
 	return int(count)
 }
